==> 00.csv_to_parquet.ipynb <==
from pyspark.sql import functions as F
from pyspark.sql import SparkSession
    
import warnings
warnings.filterwarnings("ignore")

import os
exec(open(os.path.join(os.environ["SPARK_HOME"], 'python/pyspark/shell.py')).read())

sc

df = spark.read.csv('./dataset/2019-Nov.csv', header=True, inferSchema=True)
# oct_data = spark.read.csv('./dataset/2019-Oct.csv', header=True, inferSchema=True)
# df=nov_data.unionAll(oct_data)
print((df.count(), len(df.columns)))

df.show(1, vertical = True, truncate = False)

dfs = df.withColumn('event_time',F.to_timestamp(F.regexp_replace(F.col('event_time'),' UTC','')))
dfs.show(1,vertical = True, truncate = False)

dfs.dtypes

dfs.write.mode('overwrite').parquet("./dataset/dataset.parquet") #.repartition(1)
