==> 02.purchase_behavior_feature_engineering.ipynb <==
from pyspark.sql import functions as F

from IPython.display import Markdown, display
def printmd(string, color=None):
    string = '**' + string.replace('_',' ').title() + '**'
    display(Markdown(string))
    
import pandas as pd
pd.set_option('display.max_rows',None)

import matplotlib.pyplot as plt
import seaborn as sns
sns.set(style="whitegrid", palette="pastel")

from functools import reduce

sc

df = spark.read.parquet("./dataset/dataset.parquet")

df.createOrReplaceTempView("cus_behaviors")

sql_script = \
"""
    SELECT *
    FROM cus_behaviors
    LIMIT 2
"""
spark.sql(sql_script).toPandas()

sql_script = \
"""
    SELECT 
        user_session,
        event_type,
        COUNT(event_type) AS count
    FROM cus_behaviors
    GROUP BY user_session, event_type
    ORDER BY event_type DESC, count
    ;
"""
temp = spark.sql(sql_script).groupBy('user_session').pivot('event_type').sum('count').alias('count').fillna(0)
temp.coalesce(1).write.mode('overwrite').save('./temp/num_action_by_session.parquet', format='parquet') 

num_action_by_session = pd.read_parquet('./temp/num_action_by_session.parquet',engine='pyarrow')
print(num_action_by_session.shape)
display(num_action_by_session.head(20))

sql_script = \
"""
    SELECT 
        user_session,
        MIN(event_time) AS event_time_first,
        MAX(event_time) AS event_time_last
    FROM cus_behaviors
    GROUP BY user_session
    ;
"""
temp = spark.sql(sql_script).withColumn('duration', F.col('event_time_last').cast('long') - F.col('event_time_first').cast('long'))
temp.coalesce(1).write.mode('overwrite').save('./temp/duration.parquet', format='parquet') 

duration = pd.read_parquet('./temp/duration.parquet',engine='pyarrow')
print(duration.shape)
display(duration.head(20))

sql_script = \
"""
    SELECT
        user_session,
        COUNT(DISTINCT category_id) AS num_distinct_category
    FROM cus_behaviors
    GROUP BY user_session
"""

temp = spark.sql(sql_script).fillna(0)
temp.coalesce(1).write.mode('overwrite').save('./temp/distinct_category.parquet', format='parquet') 

distinct_category = pd.read_parquet('./temp/distinct_category.parquet',engine='pyarrow')
print(distinct_category.shape)
display(distinct_category.sort_values('num_distinct_category',ascending = False).head(20))

sql_script = \
"""
    SELECT
        user_session,
        COUNT(DISTINCT category_id) AS num_distinct_cart_cat,
        COUNT(DISTINCT product_id) AS num_distinct_cart_product
    FROM cus_behaviors
    WHERE event_type = 'cart'
    GROUP BY user_session
"""

temp = spark.sql(sql_script).fillna(0)
temp.coalesce(1).write.mode('overwrite').save('./temp/num_distinct_cart_cat.parquet', format='parquet') 

num_distinct_cart_cat = pd.read_parquet('./temp/num_distinct_cart_cat.parquet',engine='pyarrow')
print(num_distinct_cart_cat.shape)
display(num_distinct_cart_cat.sort_values('num_distinct_cart_cat',ascending = False).head(20))

sql_script = \
"""
    SELECT 
        user_session,
        event_type,
        SUM(price) AS potential_revenue
    FROM cus_behaviors
    GROUP BY user_session, event_type
    ORDER BY event_type DESC, potential_revenue
    ;
"""
temp = spark.sql(sql_script).groupBy('user_session').pivot('event_type').sum('potential_revenue').alias('potential_revenue').fillna(0)
temp.coalesce(1).write.mode('overwrite').save('./temp/potential_revenue.parquet', format='parquet') 

potential_revenue = pd.read_parquet('./temp/potential_revenue.parquet',engine='pyarrow')
print(potential_revenue.shape)
display(potential_revenue.head(20)) 

num_action_by_session = pd.read_parquet('./temp/num_action_by_session.parquet',engine='pyarrow')
duration = pd.read_parquet('./temp/duration.parquet',engine='pyarrow')
distinct_category = pd.read_parquet('./temp/distinct_category.parquet',engine='pyarrow')
num_distinct_cart_cat = pd.read_parquet('./temp/num_distinct_cart_cat.parquet',engine='pyarrow')
num_action_by_session.columns = [num_action_by_session.columns[0]] + ['c_'+x for x in num_action_by_session.columns[1:]]
potential_revenue = pd.read_parquet('./temp/potential_revenue.parquet',engine='pyarrow')
potential_revenue.columns = [potential_revenue.columns[0]] + ['r_'+x for x in potential_revenue.columns[1:]]


tab_names = ['duration','distinct_category','num_distinct_cart_cat','num_action_by_session','potential_revenue']
tab_list = [eval(tab) for tab in tab_names]
df = reduce(lambda df1,df2: pd.merge(df1,df2,left_on='user_session',right_on='user_session',how = 'outer'),tab_list)

df.shape

df.fillna(0)
df.dropna()
df.shape
df.to_parquet('./dataset/temp.parquet',engine='pyarrow')

df = spark.read.parquet('./dataset/temp.parquet').drop('__index_level_0__')

from pyspark.sql import functions as F
df = df.withColumn('r_view',F.round(F.col('r_view'),2))
df = df.withColumn('mean_cart',F.round(F.col('r_cart')/F.col('c_cart'),2)).fillna(0)
df = df.withColumn('mean_purchase',F.round(F.col('r_purchase')/F.col('c_purchase'),2)).fillna(0)
df = df.withColumn('mean_view',F.round(F.col('r_view')/F.col('c_view'),2)).fillna(0)
df = df.withColumn("hour", F.hour(F.col("event_time_first")))
df = df.withColumn("day_name", F.date_format("event_time_first",'E'))

df.select(['user_session','mean_cart','mean_purchase','mean_view','hour','day_name']).show(2,truncate=False)

df.write.mode('overwrite').parquet('./dataset/refined_feature_dataset.parquet')
