==> 04.early_purchase_prediction.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.preprocessing import OneHotEncoder, MinMaxScaler
from sklearn.model_selection import train_test_split, GridSearchCV
from xgboost import XGBClassifier

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix

df = spark.read.parquet('./dataset/refined_feature_dataset.parquet')
df.createOrReplaceTempView('session_insight')

df.columns

sql_script = \
"""
    SELECT
        ROUND(duration/c_view,2) AS avg_duration,
        mean_cart,
        day_name,
        hour,
        num_distinct_category,
        num_distinct_cart_cat,
        num_distinct_cart_product,
        (CASE
            WHEN c_purchase <> 0 then 1
            ELSE c_purchase
        END) AS labels
    FROM session_insight
    WHERE user_session IS NOT NULL
    AND c_cart <> 0
    ;
"""

dataset = spark.sql(sql_script)
dataset.show(1,vertical =True)
dataset.coalesce(1).write.mode('overwrite').save('./temp/dataset.parquet', format='parquet') 

dataset = pd.read_parquet('./temp/dataset.parquet',engine='pyarrow').fillna(0)
dataset.info(show_counts=True)

print("number of purchase action: ",sum(dataset['labels'].values.tolist()))

scaler = MinMaxScaler()

scaler.fit(dataset[dataset.columns[:2]])
scaled_numeric_cols = pd.DataFrame(\
                               scaler.transform(dataset[dataset.columns[:2]]),\
                               columns=['scaled_'+col for col in dataset.columns[:2]])
scaled_numeric_cols.head()

encoder = OneHotEncoder()

encoder.fit(dataset[dataset.columns[2:4]])
encoded_categories = pd.DataFrame(\
                             encoder.transform(dataset[dataset.columns[2:4]]).toarray(),
                             columns = encoder.get_feature_names_out(dataset.columns[2:4]))
encoded_categories.head()

X = pd.merge(scaled_numeric_cols,encoded_categories,left_index=True,right_index=True,how='inner')
X[['num_distinct_category','num_distinct_cart_cat','num_distinct_cart_product']] = dataset[['num_distinct_category','num_distinct_cart_cat','num_distinct_cart_product']]
y = dataset['labels']
del scaled_numeric_cols,encoded_categories
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1, test_size=0.3, stratify=y)

clf = XGBClassifier(
    eval_metric='logloss',
    learning_rate = 0.1,
    max_depth=12,
    reg_lambda=6,
    n_estimators=100,
    use_label_encoder=False,
    tree_method='gpu_hist',
    predictor='gpu_predictor')
clf.fit(X_train, y_train)

y_hat_train = clf.predict(X_train)
y_hat_test = clf.predict(X_test)

print("train accuracy: ",  round(100*accuracy_score(y_hat_train,y_train),2),'%')
print("test accuracy: ", round(100*accuracy_score(y_hat_test,y_test),2),'%')
print("train precision: ", round(100*precision_score(y_hat_train,y_train),2),'%')
print("test precision: ", round(100*precision_score(y_hat_test,y_test),2),'%')
print("train recall: ", round(100*recall_score(y_hat_train,y_train),2),'%')
print("test recall: ", round(100*recall_score(y_hat_test,y_test),2),'%')
print("train f1: ", round(100*f1_score(y_hat_train,y_train),2),'%')
print("test f1: ", round(100*f1_score(y_hat_test,y_test),2),'%')
print('train cm')
print(confusion_matrix(y_hat_train,y_train))
print('test cm')
print(confusion_matrix(y_hat_test,y_test))

y_hat = clf.predict(X)
print(confusion_matrix(y_hat,y))
